🚚 Move bigint arbitraries into /arbitrary
