+++ src/check/arbitrary/MixedCaseArbitrary.ts
@@ -1,6 +1,6 @@
 import { Random } from '../../random/generator/Random';
 import { Stream } from '../../stream/Stream';
-import { bigUintN } from './BigIntArbitrary';
+import { bigUintN } from '../../arbitrary/bigUintN';
 import { Arbitrary } from './definition/Arbitrary';
 import { Shrinkable } from './definition/Shrinkable';
