+++ src/fast-check-default.ts
@@ -20,14 +20,10 @@ import { assert, check } from './check/runner/Runner';
 import { sample, statistics } from './check/runner/Sampler';
 
 import { array, ArrayConstraints } from './check/arbitrary/ArrayArbitrary';
-import {
-  bigInt,
-  bigIntN,
-  bigUint,
-  bigUintN,
-  BigIntConstraints,
-  BigUintConstraints,
-} from './check/arbitrary/BigIntArbitrary';
+import { bigInt, BigIntConstraints } from './arbitrary/bigInt';
+import { bigIntN } from './arbitrary/bigIntN';
+import { bigUint, BigUintConstraints } from './arbitrary/bigUint';
+import { bigUintN } from './arbitrary/bigUintN';
 import { boolean } from './check/arbitrary/BooleanArbitrary';
 import { falsy, FalsyContraints, FalsyValue } from './check/arbitrary/FalsyArbitrary';
 import { ascii, base64, char, char16bits, fullUnicode, hexa, unicode } from './check/arbitrary/CharacterArbitrary';
